design: 디테일 하단 버튼 padding 추가
